#167620
The modification was suggested by IRL.
Rule  SALE-L02-00-0045 says that the organisation is mandatory for a TOD and a specific role.
Rule SALE-L00-00-0095 is about the name of the organisation that must be present if the organisation is provided. The comment was to recall that it is in the context of TOD but it is enforce by the previous rule.
The suppression of the comment is more a correction of the rule rather than a different definition. So we keep the rule id.

Kind regards,
Thierry REMY